Refactor functions related to param_shapes and param_types
