+++ algorithmic_efficiency/workloads/cifar/workload.py
@@ -5,11 +5,6 @@ from algorithmic_efficiency import spec
 
 class BaseCifarWorkload(spec.Workload):
 
-  def __init__(self):
-    self._eval_iters = {}
-    self._param_shapes = None
-    self._param_types = None
-
   def has_reached_goal(self, eval_result: float) -> bool:
     return eval_result['validation/accuracy'] > self.target_value
 
@@ -65,22 +60,6 @@ class BaseCifarWorkload(spec.Workload):
   @property
   def eval_period_time_sec(self):
     return 600  # 10 mins
-
-  @property
-  def param_shapes(self):
-    """The shapes of the parameters in the workload model."""
-    if self._param_shapes is None:
-      raise ValueError(
-          'This should not happen, workload.init_model_fn() should be called '
-          'before workload.param_shapes!')
-    return self._param_shapes
-
-  @property
-  def model_params_types(self):
-    """
-    TODO: return shape tuples from model as a tree
-    """
-    raise NotImplementedError
 
   # Return whether or not a key in spec.ParameterTree is the output layer
   # parameters.
