+++ algorithmic_efficiency/workloads/fastmri/fastmri_jax/workload.py
@@ -19,19 +19,12 @@ from algorithmic_efficiency.workloads.fastmri.workload import \
 
 class FastMRIWorkload(BaseFastMRIWorkload):
 
-  @property
-  def model_params_types(self):
-    """The shapes of the parameters in the workload model."""
-    if self._param_types is None:
-      self._param_types = param_utils.jax_param_types(self._param_shapes)
-    return self._param_types
-
   def init_model_fn(self, rng: spec.RandomState) -> spec.ModelInitState:
     fake_batch = jnp.zeros((13, 320, 320))
     variables = jax.jit(models.UNet().init)({'params': rng}, fake_batch)
     params = variables['params']
-    self._param_shapes = jax.tree_map(lambda x: spec.ShapeTuple(x.shape),
-                                      params)
+    self._param_shapes = param_utils.jax_param_shapes(params)
+    self._param_types = param_utils.jax_param_types(self._param_shapes)
     params = jax_utils.replicate(params)
     return params, None
