+++ algorithmic_efficiency/workloads/imagenet_resnet/imagenet_pytorch/workload.py
@@ -44,13 +44,6 @@ def imagenet_v2_to_torch(batch):
 
 class ImagenetResNetWorkload(BaseImagenetResNetWorkload):
 
-  @property
-  def model_params_types(self):
-    """The shapes of the parameters in the workload model."""
-    if self._param_types is None:
-      self._param_types = param_utils.pytorch_param_types(self._param_shapes)
-    return self._param_types
-
   def _build_dataset(self,
                      data_rng: spec.RandomState,
                      split: str,
@@ -135,9 +128,8 @@ class ImagenetResNetWorkload(BaseImagenetResNetWorkload):
   def init_model_fn(self, rng: spec.RandomState) -> spec.ModelInitState:
     torch.random.manual_seed(rng[0])
     model = resnet50()
-    self._param_shapes = {
-        k: spec.ShapeTuple(v.shape) for k, v in model.named_parameters()
-    }
+    self._param_shapes = param_utils.pytorch_param_shapes(model)
+    self._param_types = param_utils.pytorch_param_types(self._param_shapes)
     model.to(DEVICE)
     if N_GPUS > 1:
       if USE_PYTORCH_DDP:
