+++ algorithmic_efficiency/workloads/librispeech_conformer/librispeech_jax/workload.py
@@ -37,9 +37,8 @@ class LibriSpeechConformerWorkload(workload.BaseLibrispeechWorkload):
                               *fake_input_batch)
 
     model_state, params = variables.pop('params')
-
-    self._param_shapes = jax.tree_map(lambda x: spec.ShapeTuple(x.shape),
-                                      params)
+    self._param_shapes = param_utils.jax_param_shapes(params)
+    self._param_types = param_utils.jax_param_types(self._param_shapes)
     model_state = jax_utils.replicate(model_state)
     params = jax_utils.replicate(params)
     return params, model_state
@@ -106,17 +105,6 @@ class LibriSpeechConformerWorkload(workload.BaseLibrispeechWorkload):
           train=False,
           mutable=False)
       return (logits, logit_paddings), None
-
-  @property
-  def model_params_types(self):
-    if self._param_shapes is None:
-      raise ValueError(
-          'This should not happen, workload.init_model_fn() should be called '
-          'before workload.param_shapes!')
-    if self._param_types is None:
-      self._param_types = param_utils.jax_param_types(
-          self._param_shapes.unfreeze())
-    return self._param_types
 
   def loss_fn(self,
               label_batch: Tuple[spec.Tensor, spec.Tensor],
