+++ algorithmic_efficiency/workloads/mnist/workload.py
@@ -17,11 +17,6 @@ USE_PYTORCH_DDP = 'LOCAL_RANK' in os.environ
 
 class BaseMnistWorkload(spec.Workload):
 
-  def __init__(self):
-    self._eval_iters = {}
-    self._param_shapes = None
-    self._param_types = None
-
   def has_reached_goal(self, eval_result: float) -> bool:
     return eval_result['validation/accuracy'] > self.target_value
 
@@ -64,15 +59,6 @@ class BaseMnistWorkload(spec.Workload):
   @property
   def eval_period_time_sec(self):
     return 10
-
-  @property
-  def param_shapes(self):
-    """The shapes of the parameters in the workload model."""
-    if self._param_shapes is None:
-      raise ValueError(
-          'This should not happen, workload.init_model_fn() should be called '
-          'before workload.param_shapes!')
-    return self._param_shapes
 
   def _eval_model(
       self,
