+++ algorithmic_efficiency/workloads/ogbg/workload.py
@@ -15,9 +15,6 @@ FLAGS = flags.FLAGS
 class BaseOgbgWorkload(spec.Workload):
 
   def __init__(self) -> None:
-    self._eval_iters = {}
-    self._param_shapes = None
-    self._param_types = None
     self._num_outputs = 128
 
   def has_reached_goal(self, eval_result: float) -> bool:
@@ -64,14 +61,6 @@ class BaseOgbgWorkload(spec.Workload):
   @property
   def eval_period_time_sec(self):
     return 120
-
-  @property
-  def param_shapes(self):
-    if self._param_shapes is None:
-      raise ValueError(
-          'This should not happen, workload.init_model_fn() should be called '
-          'before workload.param_shapes!')
-    return self._param_shapes
 
   def _build_input_queue(self,
                          data_rng: jax.random.PRNGKey,
