+++ algorithmic_efficiency/workloads/wmt/wmt_pytorch/workload.py
@@ -145,9 +145,8 @@ class WmtWorkload(BaseWmtWorkload):
   def init_model_fn(self, rng: spec.RandomState) -> spec.ModelInitState:
     torch.random.manual_seed(rng[0])
     model = Transformer()
-    self._param_shapes = {
-        k: spec.ShapeTuple(v.shape) for k, v in model.named_parameters()
-    }
+    self._param_shapes = param_utils.pytorch_param_shapes(model)
+    self._param_types = param_utils.pytorch_param_types(self._param_shapes)
     model.to(DEVICE)
     if N_GPUS > 1:
       if USE_PYTORCH_DDP:
@@ -281,13 +280,3 @@ class WmtWorkload(BaseWmtWorkload):
         aux_dropout_rate=0.1,  # Unused for eval.
         update_batch_norm=False)
     return self.compute_summed_metrics(logits, targets, weights)
-
-  @property
-  def model_params_types(self):
-    if self._param_shapes is None:
-      raise ValueError(
-          'This should not happen, workload.init_model_fn() should be called '
-          'before workload.param_shapes!')
-    if self._param_types is None:
-      self._param_types = param_utils.jax_param_types(self._param_shapes)
-    return self._param_types
